Bygetpassmethod to take input and convert to passw
getpass is a useful Python module for receiving password feedback from the user. The getpass module is illustrated in the following script below.
The getpass() method is used to take the input and convert it to a password. Besides, the if statement is used to compare the input value to the given password.
If the password matches, the message “you are authenticated” will appear; otherwise, the message “you are not authenticated” will appear.
Save the script to a file called get_pass.py with the following code.