scirpt modifiesDNSHostFile works as DNS Poisioning
Batch file can have the ability to modify the DNS hosts.txt file that resides inside ‘C:\windows\system32\drivers\etc\hosts.txt’, so that it will take you to some malicious websites instead of landing you to the legitimate website. This may also be used for phishing, i.e. redirecting you to a fake website which looks exactly like the legitimate one, created by hackers and then steal credentials. But for this script you need administrative privileges.

This program creates a new entry in the hosts file, so that whenever the user attempts to get that site he/she will be redirected to fake phishing site hosted on that IP we have changed. In the above case whenever the user tries to open www.paypal.com will be redirected to 10.115.46.73 hacker will build a fake PayPal site which looks exactly same as the original legitimate site. Such that he can get the credentials when you entered.