script prints a calendar specified monthAndYear
A program to print a calendar of the specified month and year
Python has a built-in module, the calendar, which helps to access a record of the calendar. There are many methods in this module, in this program, we are trying to print the calendar of the specified month of a given year.