script to display an alert box with message
A program to display an alert box with a message.
This program, pyautogui is used for displaying the alert box. It can be installed by using pip install pyautogui. This module has many methods and it is also commonly used to control the mouse and keyboard with a python program.