script to generate certificates using template
Automatic-Certificate-Generator
This python script automatically generates certificate by using a certificate template and csv file which contains the list of names to be printed on certificate. It downloads the certificate in the directory i.e. in pictures folder where the scripts are there.

Installation
First of all install python on your system.

pip install PI
pip install Pandas
pip install pillow
Procedure
Go to this Reference and follow it.