scriptOpenMarkdownViewerWhereUserWriteMarkdownFile
For Reference
https://github.com/avinashkranjan/Amazing-Python-Scripts/tree/master/Markdown-viewer

Markdown viewer
Running this Script opens up a markdown viewer where the user can write markdown files (readme files) and see the preview on a button click.

Setup instructions
In order to run this script, you need to have Python and pip installed on your system. After you're done installing Python and pip, run the following command from your terminal to install the requirements from the same folder (directory) of the project.

pip install -r requirements.txt
After satisfying all the requirements for the project, Open the terminal in the project folder and run

python viewer.py
or

python3 viewer.py
depending upon the python version. Make sure that you are running the command from the same virtual environment in which the required modules are installed.

Output
The user can enter the markdown data and see the preview as shown in the sample screenshot below

Markdown Viewer