script To find Hash of File
https://www.programiz.com/python-programming/examples/hash-file


Hash functions take an arbitrary amount of data and return a fixed-length bit string. The output of the function is called the digest message.

They are widely used in cryptography for authentication purposes. There are many hashing functions like MD5, SHA-1 etc. Refer this page to know more about hash functions in cryptography.

In this example, we will illustrate how to hash a file. We will use the SHA-1 hashing algorithm. The digest of SHA-1 is 160 bits long.

We do not feed the data from the file all at once, because some files are very large to fit in memory all at once. Breaking the file into small chunks will make the process memory efficient.


Output

633d7356947eec543c50b76a1852f92427f4dca9
In this program, we open the file in binary mode. Hash functions are available in the hashlib module. We loop till the end of the file using a while loop. On reaching the end, we get empty bytes object.

In each iteration, we only read 1024 bytes (this value can be changed according to our wish) from the file and update the hashing function.

Finally, we return the digest message in hexadecimal representation using the hexdigest() method.