script to compute all permutation of string
Python Program to Compute all the Permutation of the String

https://www.programiz.com/python-programming/examples/permutation-of-string

In this example, you will learn to compute all the permutation of the string.

To understand this example, you should have the knowledge of the following Python programming topics:

Python Strings
Python for Loop
Permutation is the method of selecting elements from a set in different ways.

For example: the number of ways in which characters from yup can be selected are yup, ypu, uyp, upy, puy, pyu, and not selecting any.

We will perform the same in the following examples.

Example 1: Using recursion


yup
ypu
uyp
upy
puy
pyu
None
In this example, recursion is used to find the permutations of a string yup.

The if condition prints string passed as argument if it is equal to the length of yub.
In each iteration of the for loop, each character of yup is stored in words.
The elements of words are swapped. In this way, we achieve all different combinations of characters.
This process continues until the maximum length is reached.