script to locate the address
https://www.airplane.dev/blog/12-useful-python-scripts-for-developers

6. Locate addresses
Locating an address can be useful if you're dealing with shipping or delivery logistics or for simple user-profiling tasks. To get started, install geocoder by running pip install geocoder in your terminal. The script below allows you to easily find the latitude and longitude coordinates for any address or to find an address from any set of coordinates: