ScriptShowNewItemCnInsertedAnyItemSearchdInHistory
Add and search data in the dictionary:
dictionary object is used in python to store multiple data like the associative array of other programming languages. The following script shows how a new item can be inserted, and any item can be searched in the dictionary. A dictionary of customer information is declared in the script where the index contains the customer ID, and the value contains the customer name. Next, one new customer information is inserted at the end of the dictionary. A customer ID is taken as input to search in the dictionary. ‘for’ loop and ‘if’ condition is used to iterate the indexes of the dictionary and search the input value in the dictionary.

c19.py


# Define a dictionary
customers = {'06753': 'Mehzabin Afroze', '02457': 'Md. Ali',
             '02834': 'Mosarof Ahmed', '05623': 'Mila Hasan', '07895': 'Yaqub Ali'}

# Append a new data
customers['05634'] = 'Mehboba Ferdous'

print("The customer names are:")
# Print the values of the dictionary
for customer in customers:
    print(customers[customer])

# Take customer ID as input to search
name = input("Enter customer ID:")

# Search the ID in the dictionary
for customer in customers:
    if customer == name:
        print(customers[customer])
        break
The following output will appear after executing the script and taking ‘02457’ as ID value.