script performs databasebackup&restoresUsingSQLite
SQLite is a C library that provides a lightweight disk-based database that doesn’t require a separate server process and allows accessing the database using a nonstandard variant of the SQL query language. Some applications can use SQLite for internal data storage. It’s also possible to prototype an application using SQLite and then port the code to a larger database such as PostgreSQL or Oracle.

In this code:

backup_database() function makes a copy of the source SQLite database file and names it as the backup file. You can run this function to create backup of your database.
restore_database() function copies the backup file back to the source file, effectively restoring the database to the state it was when the backup was created.
The user is presented with options to either backup the database, restore it, or quit the program.
You can adjust the source_db_file and backup_db_file variables to specify the paths to your SQLite source and backup database files